pyproject.toml
[tool.poetry]
name = "lunchlog"
version = "0.1.0"
description = "A lunch logging application"
authors = ["karippery <john.karippery@yahoo.com>"]
readme = "README.md"
packages = [{include = "lunchlog"}]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
drf-yasg = "^1.21.10"
django-extensions = "^4.1"
django-storages = "^1.14.6"
boto3 = "^1.39.11"
pillow = "^11.3.0"
botocore = "^1.39.11"



[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"
django-debug-toolbar = "^4.3.0"
pytest = "^8.2"
pytest-django = "^4.8"
factory-boy = "^3.3"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py", "*_test.py"]